Move status widget to separate file and add widget style
