+++ src/style/StatusWidget.ts
@@ -1,0 +1,5 @@
+import { style } from 'typestyle';
+
+export const statusWidgetClass = style({
+  lineHeight: '24px'
+});
